requirements.txt
six==1.9.0
yolk==0.4.3
Django==1.8.0
PyYAML==3.10
chardet==2.1.1
dj-database-url==0.2.2
dj-static==0.0.5
django-admin-import==0.3.0
django-appconf==0.6
gunicorn==17.5
psycopg2==2.5.1
static==0.4
wsgiref==0.1.2
nltk==2.0.4
patsy==0.2.1
django-picklefield==0.3.0
django-nose==1.2
django-easydump==0.2.6
django-dajax==0.9.2
django-dajaxice==0.7
newrelic==2.16.0.12
django-extensions==1.5.5
textblob==0.9.0

# Note: All these require numpy to be installed first
pandas==0.11.0

#On heroku, this requires pandas to be installed first
scikit-learn==0.14.1


